Unified logic for displaying and exporting results to use the same method https://github.com/SoftUni-Internal/suls-issues/issues/4211
