+++ Open Judge System/Web/OJS.Web/App_GlobalResources/Areas/Administration/Contests/ContestsControllers.resx
@@ -129,9 +129,6 @@
   <data name="Added_time_to_participants_online" xml:space="preserve">
     <value>Successfully added {0} minutes to the times of all active participants in the contest {1}</value>
   </data>
-  <data name="Contest" xml:space="preserve">
-    <value>contest</value>
-  </data>
   <data name="Contest_added" xml:space="preserve">
     <value>Contest added successfully</value>
   </data>
@@ -168,17 +165,11 @@
   <data name="Participants_transferred" xml:space="preserve">
     <value>Participants transferred successfully</value>
   </data>
-  <data name="Practice" xml:space="preserve">
-    <value>practice</value>
-  </data>
   <data name="Practice_start_date_before_end" xml:space="preserve">
     <value>Practice start date must be before the practice end date</value>
   </data>
   <data name="Problem_groups_count_limit" xml:space="preserve">
     <value>Problem groups count cannot be more than {0}</value>
-  </data>
-  <data name="Report_excel_format" xml:space="preserve">
-    <value>Ranking for {0} {1}.xls</value>
   </data>
   <data name="Report_zip_format" xml:space="preserve">
     <value>{1} submissions for {0}.zip</value>
