+++ Open Judge System/Web/OJS.Web/App_GlobalResources/Areas/Contests/ContestsGeneral.resx
@@ -195,4 +195,13 @@
   <data name="Problem_not_assigned_to_user" xml:space="preserve">
     <value>The problem is not part of your problems for the exam</value>
   </data>
+  <data name="Contest" xml:space="preserve">
+    <value>Contest</value>
+  </data>
+  <data name="Practice" xml:space="preserve">
+    <value>Practice</value>
+  </data>
+  <data name="Report_excel_format" xml:space="preserve">
+    <value>Ranking for {0} {1}.xls</value>
+  </data>
 </root>
