+++ Open Judge System/Web/OJS.Web/Areas/Administration/Controllers/ContestsExportController.cs
@@ -9,9 +9,6 @@
 
     using Ionic.Zip;
 
-    using NPOI.HSSF.UserModel;
-    using NPOI.SS.UserModel;
-
     using OJS.Common;
     using OJS.Common.Extensions;
     using OJS.Common.Models;
@@ -23,164 +20,11 @@
     using OJS.Web.Common;
     using OJS.Web.Common.Extensions;
 
-    using Resource = Resources.Areas.Administration.Contests.ContestsControllers;
-
     public class ContestsExportController : LecturerBaseController
     {
         public ContestsExportController(IOjsData data)
             : base(data)
         {
-        }
-
-        public ActionResult Results(int id, bool compete)
-        {
-            var userHasAccessToContest =
-                this.User.IsAdmin() ||
-                this.Data.Contests
-                    .All()
-                    .Any(c =>
-                        c.Id == id &&
-                        (c.Lecturers.Any(cl => cl.LecturerId == this.UserProfile.Id) ||
-                            c.Category.Lecturers.Any(cl => cl.LecturerId == this.UserProfile.Id)));
-            if (!userHasAccessToContest)
-            {
-                this.TempData.AddDangerMessage("Нямате привилегиите за това действие");
-                return this.RedirectToAction("Index", "Contests");
-            }
-
-            var contest = this.Data.Contests.GetById(id);
-
-            var data = new
-            {
-                contest.Id,
-                contest.Name,
-                Problems = contest.ProblemGroups
-                    .Where(pg => !pg.IsDeleted)
-                    .SelectMany(pg => pg.Problems.Where(p => !p.IsDeleted))
-                    .AsQueryable()
-                    .OrderBy(x => x.OrderBy)
-                    .ThenBy(x => x.Name),
-                Questions = contest.Questions.OrderBy(x => x.Id),
-                Results = this.Data.Participants.All()
-                    .Where(participant => participant.ContestId == contest.Id && participant.IsOfficial == compete)
-                    .Select(participant => new
-                    {
-                        ParticipantUserName = participant.User.UserName,
-                        ParticipantFirstName = participant.User.UserSettings.FirstName,
-                        ParticipantLastName = participant.User.UserSettings.LastName,
-                        Answers = participant.Answers.OrderBy(answer => answer.ContestQuestionId),
-                        ProblemResults = participant.Contest.ProblemGroups
-                            .Where(pg => !pg.IsDeleted)
-                            .SelectMany(pg => pg.Problems.Where(p => !p.IsDeleted))
-                            .Select(problem =>
-                                new
-                                {
-                                    problem.Id,
-                                    ProblemName = problem.Name,
-                                    ProblemOrderBy = problem.OrderBy,
-                                    ShowResult = problem.ShowResults,
-                                    BestSubmission = problem.Submissions
-                                                        .Where(z => z.ParticipantId == participant.Id)
-                                                        .OrderByDescending(z => z.Points).ThenByDescending(z => z.Id)
-                                                        .Select(z => new { z.Id, z.Points })
-                                                        .FirstOrDefault()
-                                })
-                                .OrderBy(res => res.ProblemOrderBy).ThenBy(res => res.ProblemName),
-                    })
-                    .ToList()
-                    .Select(x => new { Data = x, Total = x.ProblemResults.Where(y => y.ShowResult).Sum(y => y.BestSubmission == null ? 0 : y.BestSubmission.Points) })
-                    .OrderByDescending(x => x.Total)
-            };
-
-            var workbook = new HSSFWorkbook();
-            var sheet = workbook.CreateSheet();
-
-            // Header
-            var headerRow = sheet.CreateRow(0);
-            int columnNumber = 0;
-            headerRow.CreateCell(columnNumber++).SetCellValue("Username");
-            headerRow.CreateCell(columnNumber++).SetCellValue("Name");
-            foreach (var question in data.Questions)
-            {
-                headerRow.CreateCell(columnNumber++).SetCellValue(question.Text);
-            }
-
-            foreach (var problem in data.Problems)
-            {
-                headerRow.CreateCell(columnNumber++).SetCellValue(problem.Name);
-            }
-
-            var totalPointsCellTitle = "Total";
-
-            if (contest.IsOnline)
-            {
-                var maxPointsForOnlineContest = contest.ProblemGroups
-                    .Where(pg => pg.Problems.Any(p => !p.IsDeleted))
-                    .Sum(pg => pg.Problems.First().MaximumPoints);
-
-                totalPointsCellTitle = $"{totalPointsCellTitle} (Max: {maxPointsForOnlineContest})";
-            }
-
-            headerRow.CreateCell(columnNumber++).SetCellValue(totalPointsCellTitle);
-
-            // All rows
-            var rowNumber = 1;
-            foreach (var result in data.Results)
-            {
-                var cellNumber = 0;
-                var row = sheet.CreateRow(rowNumber++);
-                row.CreateCell(cellNumber++).SetCellValue(result.Data.ParticipantUserName);
-                row.CreateCell(cellNumber++).SetCellValue(
-                    $"{result.Data.ParticipantFirstName} {result.Data.ParticipantLastName}".Trim());
-                foreach (var answer in result.Data.Answers)
-                {
-                    int answerId;
-                    if (answer.ContestQuestion.Type == ContestQuestionType.DropDown && int.TryParse(answer.Answer, out answerId))
-                    {
-                        // TODO: N+1 query problem. Optimize it.
-                        var answerText =
-                            this.Data.ContestQuestionAnswers.All()
-                                .Where(x => x.Id == answerId)
-                                .Select(x => x.Text)
-                                .FirstOrDefault();
-                        row.CreateCell(cellNumber++).SetCellValue(answerText);
-                    }
-                    else
-                    {
-                        row.CreateCell(cellNumber++).SetCellValue(answer.Answer);
-                    }
-                }
-
-                foreach (var problemResult in result.Data.ProblemResults)
-                {
-                    if (problemResult.BestSubmission != null)
-                    {
-                        row.CreateCell(cellNumber++).SetCellValue(problemResult.BestSubmission.Points);
-                    }
-                    else
-                    {
-                        row.CreateCell(cellNumber++, CellType.Blank);
-                    }
-                }
-
-                row.CreateCell(cellNumber++).SetCellValue(result.Total);
-            }
-
-            // Auto-size all columns
-            for (var i = 0; i < columnNumber; i++)
-            {
-                sheet.AutoSizeColumn(i);
-            }
-
-            // Write the workbook to a memory stream
-            var outputStream = new MemoryStream();
-            workbook.Write(outputStream);
-
-            // Return the result to the end user
-            return this.File(
-                outputStream.ToArray(), // The binary data of the XLS file
-                GlobalConstants.ExcelMimeType, // MIME type of Excel files
-                string.Format(Resource.Report_excel_format, compete ? Resource.Contest : Resource.Practice, contest.Name)); // Suggested file name in the "Save as" dialog which will be displayed to the end user
         }
 
         public ActionResult Solutions(int id, bool compete, SubmissionExportType exportType)
